+++ truck-meshalgo/src/tessellation/mod.rs
@@ -3,100 +3,44 @@ use spade::{iterators::*, *};
 use truck_topology::{compress::*, *};
 
 #[cfg(not(target_arch = "wasm32"))]
-mod meshables_traits {
-    use super::*;
-    /// Gathered the traits used in tessellation.
-    pub trait PolylineableCurve:
-        ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3> + Send + Sync {
-    }
-    impl<
-            C: ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3> + Send + Sync,
-        > PolylineableCurve for C
-    {
-    }
-    /// Gathered the traits used in tessellation by [`MeshableShape`].
-    pub trait MeshableSurface:
-        ParametricSurface3D
-        + ParameterDivision2D
-        + SearchParameter<D2, Point = Point3>
-        + Send
-        + Sync {
-    }
-    impl<
-            S: ParametricSurface3D
-                + ParameterDivision2D
-                + SearchParameter<D2, Point = Point3>
-                + Send
-                + Sync,
-        > MeshableSurface for S
-    {
-    }
-    /// Gathered the traits used in robust tessellation by [`RobustMeshableShape`].
-    pub trait RobustMeshableSurface:
-        ParametricSurface3D
-        + ParameterDivision2D
-        + SearchParameter<D2, Point = Point3>
-        + SearchNearestParameter<D2, Point = Point3>
-        + Send
-        + Sync {
-    }
-    impl<
-            S: ParametricSurface3D
-                + ParameterDivision2D
-                + SearchParameter<D2, Point = Point3>
-                + SearchNearestParameter<D2, Point = Point3>
-                + Send
-                + Sync,
-        > RobustMeshableSurface for S
-    {
-    }
-
-    pub(super) trait PreMeshableSurface:
-        ParametricSurface3D + ParameterDivision2D + Send + Sync {
-    }
-    impl<S: ParametricSurface3D + ParameterDivision2D + Send + Sync> PreMeshableSurface for S {}
+mod parallelizable {
+    /// Parallelizable by `rayon`.
+    pub trait Parallelizable: Send + Sync {}
+    impl<T: Send + Sync> Parallelizable for T {}
 }
 
 #[cfg(target_arch = "wasm32")]
-mod meshables_traits {
-    use super::*;
-    /// Gathered the traits used in tessellation.
-    pub trait PolylineableCurve:
-        ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3> {
-    }
-    impl<C: ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3>>
-        PolylineableCurve for C
-    {
-    }
-    /// Gathered the traits used in tessellation by [`MeshableShape`].
-    pub trait MeshableSurface:
-        ParametricSurface3D + ParameterDivision2D + SearchParameter<D2, Point = Point3> {
-    }
-    impl<S: ParametricSurface3D + ParameterDivision2D + SearchParameter<D2, Point = Point3>>
-        MeshableSurface for S
-    {
-    }
-    /// Gathered the traits used in tessellation by [`RobustMeshableShape`].
-    pub trait RobustMeshableSurface:
-        ParametricSurface3D
-        + ParameterDivision2D
-        + SearchParameter<D2, Point = Point3>
-        + SearchNearestParameter<D2, Point = Point3> {
-    }
-    impl<
-            S: ParametricSurface3D
-                + ParameterDivision2D
-                + SearchParameter<D2, Point = Point3>
-                + SearchNearestParameter<D2, Point = Point3>,
-        > RobustMeshableSurface for S
-    {
-    }
-
-    pub(super) trait PreMeshableSurface: ParametricSurface3D + ParameterDivision2D {}
-    impl<S: ParametricSurface3D + ParameterDivision2D> PreMeshableSurface for S {}
+mod parallelizable {
+    /// No parallelization in the case of wasm.
+    pub trait Parallelizable {}
+    impl<T> Parallelizable for T {}
 }
 
-pub use meshables_traits::*;
+pub use parallelizable::*;
+
+/// Gathered the traits used in tessellation.
+pub trait PolylineableCurve:
+    ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3> + Parallelizable {
+}
+impl<
+        C: ParametricCurve3D + BoundedCurve + ParameterDivision1D<Point = Point3> + Parallelizable,
+    > PolylineableCurve for C
+{
+}
+
+/// It can be meshed, but not necessarily trimmed.
+pub trait PreMeshableSurface: ParametricSurface3D + ParameterDivision2D + Parallelizable {}
+impl<S: ParametricSurface3D + ParameterDivision2D + Parallelizable> PreMeshableSurface for S {}
+
+/// The generated mesh can be trimmed only if the boundary curves ride strictly on a surface.
+pub trait MeshableSurface: PreMeshableSurface + SearchParameter<D2, Point = Point3> {}
+impl<S: PreMeshableSurface + SearchParameter<D2, Point = Point3>> MeshableSurface for S {}
+
+/// The generated mesh can be trimmed if the boundary curves does not ride strictly on a surface.
+pub trait RobustMeshableSurface:
+    MeshableSurface + SearchNearestParameter<D2, Point = Point3> {
+}
+impl<S: MeshableSurface + SearchNearestParameter<D2, Point = Point3>> RobustMeshableSurface for S {}
 
 type PolylineCurve = truck_polymesh::PolylineCurve<Point3>;